+++ database/migrations/2022_09_05_181138_create_likes_table.php
@@ -33,6 +33,10 @@ return new class extends Migration
      */
     public function down(): void
     {
+        if (config('database.default') !== 'mysql_morph') {
+            return;
+        }
+
         Schema::dropIfExists('likes');
     }
 };
